requirements-dev.txt
# Development dependencies (includes test and debugging tools)
-r requirements.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
black==23.12.0
flake8==6.1.0
pylint==3.0.3

# Debugging
Flask-DebugToolbar==0.14.1

# Developer utilities
ipython==8.18.1
